requirements.txt
Pillow>=10.0.0
requests>=2.31.0
urllib3>=2.0.0 
